FIX: Fixed layout issue with XInput on OSX and Windows Desktops. (#531)
* FIX: Fixed layout issue with XInput on OSX and Windows Desktops.

* FIX: Forgot to add CHANGELOG and properly cast the buttons.

+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XInput/XInputControllerOSX.cs
@@ -65,7 +65,7 @@ namespace UnityEngine.Experimental.Input.Plugins.XInput.LowLevel
         [InputControl(name = "buttonNorth", bit = (uint)Button.Y, displayName = "Y")]
 
         [FieldOffset(2)]
-        public uint buttons;
+        public ushort buttons;
 
         [InputControl(name = "leftTrigger", format = "BYTE")]
         [FieldOffset(4)] public byte leftTrigger;
@@ -99,7 +99,7 @@ namespace UnityEngine.Experimental.Input.Plugins.XInput.LowLevel
 
         public XInputControllerOSXState WithButton(Button button)
         {
-            buttons |= (uint)1 << (int)button;
+            buttons |= (ushort)((uint)1 << (int)button);
             return this;
         }
     }
